+++ distributions/univariate/continuous/NormalRand.h
@@ -12,7 +12,7 @@
  */
 class RANDLIBSHARED_EXPORT NormalZiggurat {
 
-    static constexpr int TABLE_SIZE = 257;
+    static constexpr size_t TABLE_SIZE = 257;
     static constexpr std::array<LongDoublePair, TABLE_SIZE> createZiggurat()
     {
         constexpr long double A = 4.92867323399e-3l; /// area under rectangle
